+++ mycodo/tests/software_tests/factories_user.py
@@ -1,13 +1,13 @@
 # coding=utf-8
 """ A collection of model factories using factory boy """
 import factory
-from mycodo.databases.users_db import models
+from mycodo.databases.mycodo_db import models_5
 
 
 class UserFactory(factory.Factory):
     """ A factory for creating user models """
     class Meta(object):
-        model = models.Users
+        model = models_5.Users
 
     user_name = factory.Faker('name')
     user_email = factory.Faker('email')
@@ -17,7 +17,7 @@ class UserFactory(factory.Factory):
 class AdminFactory(factory.Factory):
     """ A factory for creating admin user models """
     class Meta(object):
-        model = models.Users
+        model = models_5.Users
 
     user_name = factory.Faker('name')
     user_email = factory.Faker('email')
@@ -28,7 +28,7 @@ class AdminFactory(factory.Factory):
 class GuestFactory(factory.Factory):
     """ A factory for creating admin user models """
     class Meta(object):
-        model = models.Users
+        model = models_5.Users
 
     user_name = factory.Faker('name')
     user_email = factory.Faker('email')